requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
django-timestampedmodel==0.1.0
factory-boy==3.2.1
Faker==18.10.1
